translate hints to en

+++ arch1.sh
@@ -6,41 +6,38 @@ echo 'Синхронизация системных часов'
 timedatectl set-ntp true
 
 
-echo '###### Разметка диска ######'
+echo '###### Disk Partitions ######'
 echo "
-Сейчас диски выглядят так:
+Right now your disks look like this:
      "
 echo " "
 fdisk -l
 echo "
-
-    Тут можно выбрать
-
-    1 -  автоматическую разметку: 4 раздела
+    1 - automatic: 4 partitions
     /dev/sda1 - boot = 100M
     /dev/sda2 - root = 30G
     /dev/sda3 - swap = 1024M
-    /dev/sda4 - home = остаток
+    /dev/sda4 - home = remaining space
     
-    0 - либо же сделать разметку вручную
-    предполагается, что будет также 3 раздела
-    отличие заключается лишь в том, что доступно изменить размеры дисков
-    иерархия для корректной работы должна оставаться такой же
-    
-    введите цифру своего предпочтения:"
+    0 - manually 
+    it must be consists of 4 partitions
+    difference with the first method in manual determining of partition's size
+    but with the same hierarchy
+  
+    input your choice:"
     echo " "
 while 
     read -n1 -p  "
-    1 - автоматически    
-    0 - вручную: " cfdisk # sends right after the keypress
+    1 - automatically    
+    0 - manually: " cfdisk # sends right after the keypress
     echo ''
     [[ "$cfdisk" =~ [^10] ]]
 do
     :
 done
  if [[ $cfdisk == 1 ]]; then
-  echo 'автоматическая разметка'
-  echo 'создание разделов'
+  echo 'automatically'
+  echo 'creating of partitions...'
   (
   echo o;
   #boot
@@ -73,13 +70,13 @@ done
   echo w;
 ) | fdisk /dev/sda
 
-  echo 'Форматирование дисков'
+  echo 'formatting of partitions...'
   mkfs.ext2  /dev/sda1 -L boot
   mkfs.ext4  /dev/sda2 -L root
   mkswap /dev/sda3 -L swap
   mkfs.ext4  /dev/sda4 -L home
 
-  echo 'Монтирование дисков'
+  echo 'mounting of partitions...'
   mount /dev/sda2 /mnt
   mkdir /mnt/{boot,home}
   mount /dev/sda1 /mnt/boot
@@ -87,41 +84,47 @@ done
   mount /dev/sda4 /mnt/home
 
 elif [[ $cfdisk == 0 ]]; then
-  read -p "Укажите диск (sda/sdb например sda) : " cfd
+  read -p "select disk (for example: sda) : " cfd
   cfdisk /dev/$cfd  
-  read -p "Укажите boot раздел (sda1/sdb1):" bootd
+  read -p "select boot partition (for example: sda1):" bootd
   mkfs.ext2 /dev/$bootd -L boot
  
     #elif [[ $boots == 0 ]]; then
     #echo "boot раздел пропущен"   
-  read -p "Укажите ROOT раздел(sda/sdb 1.2.3.4 (sda5 например)):" root
+  read -p "select ROOT partition (for example: sda2 ):" root
   echo ""
   mkfs.ext4 /dev/$root -L root
   mount /dev/$root /mnt
   mkdir /mnt/{boot,home}
   mount
   echo ""
-  ########## boot  ########
+  ########## boot ##########
   mount /dev/$bootd /mnt/boot
-  ############ swap   #########################################################
-  read -p "Укажите swap раздел(sda/sdb 1.2.3.4 (sda7 например)):" swaps
+  ############ swap ############
+  read -p "select swap partition (for example: sda3):" swaps
   swapon /dev/$swaps
-  ################  home     ############################################################ 
+  ########### home ########### 
   clear
-  read -p "Укажите HOME раздел(sda1/sdb3):" home
+  read -p "select home partition(for example: sda4):" home
   mount /dev/$home /mnt/home
 fi
 
-echo 'разметка диска'
+echo 'now:'
 fdisk -l
 
 echo 'Выбор зеркал для загрузки. Ставим зеркало от Яндекс'
-echo "Server = http://mirror.yandex.ru/archlinux/\$repo/os/\$arch" > /etc/pacman.d/mirrorlist
+read -p"if you need to set up mirror from yandex,  press 1:"mirr
+if [[$mirr == 1]]; then
+  echo "Server = http://mirror.yandex.ru/archlinux/\$repo/os/\$arch" > /etc/pacman.d/mirrorlist
+elif [[$mirr != 1]]; then
+  echo "setting up of mirror skipped"
+fi
 
-echo 'Установка основных пакетов'
+
+echo 'Installing of base packages...'
 pacstrap /mnt base base-devel linux linux-firmware nano dhcpcd netctl
 
-echo 'Настройка системы'
+echo 'Setting up of the system...'
 genfstab -pU /mnt >> /mnt/etc/fstab
 
 arch-chroot /mnt sh -c "$(curl -fsSL https://raw.githubusercontent.com/UserNameMikko/archinstall/master/arch2.sh)"
